Moved Pimple integration to the correct namespace
